+++ src/graphic/index_buffer.h
@@ -4,7 +4,7 @@ class IndexBuffer {
 private:
     unsigned int id_ = 0;
 public:
-    IndexBuffer(const void* const data, unsigned int size);
+    IndexBuffer(const void* const data, size_t size);
     void bind() const;
     void unbind() const;
 };
